WIP clean up distinction between midi manager and device
